+++ lib/puppet_check.rb
@@ -101,11 +101,11 @@ class PuppetCheck
     paths.uniq.each do |path|
       if File.directory?(path)
         # glob all files in directory and concat them
-        files.concat(Dir.glob("#{path}/**/*").select { |subpath| File.file?(subpath) && !subpath.include?('fixtures') })
-      elsif File.file?(path) && !path.include?('fixtures')
+        files.concat(Dir.glob("#{path}/**/*").select { |subpath| File.readable?(subpath) && !subpath.include?('fixtures') })
+      elsif File.readable?(path) && !path.include?('fixtures')
         files.push(path)
       else
-        warn "puppet-check: #{path} is not a directory, file, or symlink, and will not be considered during parsing"
+        warn "puppet-check: #{path} is not a readable directory, file, or symlink, and will not be considered during parsing"
       end
     end